+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/wallet/account/AssetPositionProcessor.kt
@@ -4,7 +4,6 @@ import exchange.dydx.abacus.output.account.PositionSide
 import exchange.dydx.abacus.processor.base.BaseProcessor
 import exchange.dydx.abacus.protocols.ParserProtocol
 import exchange.dydx.abacus.state.internalstate.InternalAssetPositionState
-import indexer.codegen.IndexerAssetPositionResponse
 import indexer.codegen.IndexerAssetPositionResponseObject
 
 internal interface AssetPositionProcessorProtocol {
