+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/wallet/account/AssetPositionsProcessor.kt
@@ -3,30 +3,31 @@ package exchange.dydx.abacus.processor.wallet.account
 import exchange.dydx.abacus.processor.base.BaseProcessor
 import exchange.dydx.abacus.protocols.ParserProtocol
 import exchange.dydx.abacus.state.internalstate.InternalAssetPositionState
-import exchange.dydx.abacus.state.internalstate.InternalPerpetualPosition
 import exchange.dydx.abacus.utils.mutable
 import exchange.dydx.abacus.utils.safeSet
-import indexer.codegen.IndexerAssetPositionResponse
 import indexer.codegen.IndexerAssetPositionResponseObject
 
-internal class AssetPositionsProcessor(parser: ParserProtocol) : BaseProcessor(parser) {
-    private val itemProcessor = AssetPositionProcessor(parser = parser)
+internal class AssetPositionsProcessor(
+    parser: ParserProtocol,
+    private val itemProcessor: AssetPositionProcessorProtocol = AssetPositionProcessor(parser = parser)
+) : BaseProcessor(parser) {
 
     // From REST call
     fun process(
-         payload: Map<String, IndexerAssetPositionResponseObject>?
+        payload: Map<String, IndexerAssetPositionResponseObject>?
     ): Map<String, InternalAssetPositionState>? {
         return if (payload != null) {
             val result = mutableMapOf<String, InternalAssetPositionState>()
             for ((key, value) in payload) {
                 val assetPosition = itemProcessor.process(value)
-                if(assetPosition != null){
+                if (assetPosition != null) {
                     result[key] = assetPosition
                 }
             }
             result
-        } else
+        } else {
             null
+        }
     }
 
     // From Websocket
@@ -43,16 +44,18 @@ internal class AssetPositionsProcessor(parser: ParserProtocol) : BaseProcessor(p
                 }
             }
             modified
-        } else
+        } else {
             existing
+        }
     }
 
-    internal fun received(payload: Map<String, Any>?): Map<String, Any>? {
+    internal fun receivedDeprecated(payload: Map<String, Any>?): Map<String, Any>? {
         if (payload != null) {
             val result = mutableMapOf<String, Any>()
             for ((key, value) in payload) {
                 val map = parser.asNativeMap(value)
                 if (map != null) {
+                    val itemProcessor = itemProcessor as AssetPositionProcessor
                     val assetPosition = itemProcessor.received(null, map)
                     result.safeSet(key, assetPosition)
                 }
@@ -62,7 +65,7 @@ internal class AssetPositionsProcessor(parser: ParserProtocol) : BaseProcessor(p
         return null
     }
 
-    internal fun receivedChanges(
+    internal fun receivedChangesDeprecated(
         existing: Map<String, Any>?,
         payload: List<Any>?
     ): Map<String, Any>? {
@@ -72,6 +75,7 @@ internal class AssetPositionsProcessor(parser: ParserProtocol) : BaseProcessor(p
                 parser.asNativeMap(item)?.let { item ->
                     val assetId = parser.asString(item["symbol"] ?: item["denom"])
                     if (assetId != null) {
+                        val itemProcessor = itemProcessor as AssetPositionProcessor
                         val modified =
                             itemProcessor.receivedChanges(parser.asNativeMap(existing?.get(assetId)), item)
                         output.safeSet(assetId, modified)
